Added new nav mesh area type Aoe with avoidance priority
Added a navmesh obstacle with carving so enemies wont go outside the boss fight
please add these outside the other rooms and put them inside the list with the doors so that they will be turned on when you collide with them therefore no enemy will be able to leave the current room as there wont be any navmesh outside. make sure to add right layers for all the walls and obstacles. the right layer is either default or obstacle both should work this will stop the gabbar from going through walls in his charge. And as for the pc being pushed through walls I am currently trying to find a fix for it.